Kernel 438100c554107314a8c5f5e894fd9aba86c1bf2f, LibLinux 5cf0c4be0d0e6b5fc34b74eb998077f5e72482d4

+++ headers/libx/xenus_chain.h
@@ -19,7 +19,7 @@ typedef void(*_pre_chain_dealloc)(void * buffer);
     etc etc 
 */
 
-XENUS_SYM error_t chain_allocate(void * bk, uint64_t hash, uint32_t length, void ** bkout, void ** out);
+XENUS_SYM error_t chain_allocate(void * bk, uint64_t hash, size_t length, void ** bkout, void ** out);
 XENUS_SYM error_t chain_deallocate_ex(void * bk, uint64_t hash, _pre_chain_dealloc cb);
 XENUS_SYM error_t chain_deallocate(void * bk, uint64_t hash);
 XENUS_SYM error_t chain_get(void * bk, uint64_t hash, void ** bnext, void ** out);
